# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bff_calendar",
    "version": "1.0.0"
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CalendarAssignee": {
        "type": "object",
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ],
            "description": "Identifiant unique de la personne",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Nom complet affichable",
            "example": "Security Admin"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Adresse email",
            "example": "security-admin@mairie360.fr"
          },
          "role": {
            "type": "string",
            "description": "Fonction ou rôle métier",
            "example": "Admin"
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri",
            "description": "URL de l'image de profil",
            "example": "https://mairie360.fr/avatar.png"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "CalendarEvent": {
        "type": "object",
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ],
            "description": "Identifiant stable. Optionnel en création, obligatoire en lecture."
          },
          "title": {
            "type": "string",
            "pattern": "^[^<>]*$",
            "description": "Titre de l'événement",
            "example": "Scan event"
          },
          "date": {
            "type": "string",
            "pattern": "^(?:(\\d{4})-(\\d{2})-(\\d{2})|(\\d{2})[-/](\\d{2})[-/](\\d{4}))$",
            "description": "Date de début (format YYYY-MM-DD ou DD-MM-YYYY)",
            "example": "2030-06-15"
          },
          "endDate": {
            "type": "string",
            "pattern": "^(?:(\\d{4})-(\\d{2})-(\\d{2})|(\\d{2})[-/](\\d{2})[-/](\\d{4}))$",
            "description": "Date de fin pour les événements multi-jours (format YYYY-MM-DD ou DD-MM-YYYY)",
            "example": "2030-06-15"
          },
          "category": {
            "type": "string",
            "enum": [
              "meeting",
              "activity",
              "ceremony",
              "other"
            ],
            "description": "Catégorie: meeting (Réunion), activity (Animation), ceremony (Cérémonie), other (Autre)"
          },
          "service": {
            "type": "string",
            "pattern": "^[^<>]*$",
            "description": "Service ou département organisateur",
            "example": "direction"
          },
          "startTime": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "description": "Heure de début au format HH:mm",
            "example": "09:00"
          },
          "endTime": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "description": "Heure de fin au format HH:mm",
            "example": "10:00"
          },
          "location": {
            "type": "string",
            "pattern": "^[^<>]*$",
            "description": "Lieu de l'événement",
            "example": "Town hall"
          },
          "description": {
            "type": "string",
            "pattern": "^[^<>]*$",
            "description": "Description détaillée"
          },
          "assigneeIds": {
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ],
              "example": 1
            },
            "description": "Liste des identifiants des personnes assignées"
          },
          "assignees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CalendarAssignee"
            },
            "description": "Objets complets des personnes assignées (optionnel, peut être reconstruit côté front)"
          },
          "recurrence": {
            "$ref": "#/components/schemas/CalendarRecurrence"
          },
          "approvalStatus": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ],
            "description": "Statut d'approbation de l'événement"
          },
          "createdById": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ],
            "description": "Identifiant de l'utilisateur ayant créé l'événement",
            "example": 1
          },
          "canValidate": {
            "type": "boolean",
            "description": "Indique si l’utilisateur courant peut valider ou refuser cet événement"
          },
          "canEdit": {
            "type": "boolean",
            "description": "Indique si l’utilisateur courant peut modifier cet événement"
          },
          "canDelete": {
            "type": "boolean",
            "description": "Indique si l’utilisateur courant peut supprimer cet événement"
          },
          "visibleToRoles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "user",
                "responsable",
                "mayor"
              ]
            },
            "description": "Liste des rôles pouvant voir l'événement"
          }
        },
        "required": [
          "title",
          "date"
        ]
      },
...
      "CreateCalendarEventBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "pattern": "^[^<>]*$",
            "description": "Titre de l'événement",
            "example": "Scan event"
          },
          "date": {
            "type": "string",
            "pattern": "^(?:(\\d{4})-(\\d{2})-(\\d{2})|(\\d{2})[-/](\\d{2})[-/](\\d{4}))$",
            "description": "Date de début (format YYYY-MM-DD ou DD-MM-YYYY)",
            "example": "2030-06-15"
          },
          "endDate": {
            "type": "string",
            "pattern": "^(?:(\\d{4})-(\\d{2})-(\\d{2})|(\\d{2})[-/](\\d{2})[-/](\\d{4}))$",
            "description": "Date de fin pour les événements multi-jours (format YYYY-MM-DD ou DD-MM-YYYY)",
            "example": "2030-06-15"
          },
          "category": {
            "type": "string",
            "enum": [
              "meeting",
              "activity",
              "ceremony",
              "other"
            ],
            "description": "Catégorie: meeting (Réunion), activity (Animation), ceremony (Cérémonie), other (Autre)"
          },
          "service": {
            "type": "string",
            "pattern": "^[^<>]*$",
            "description": "Service ou département organisateur",
            "example": "direction"
          },
          "startTime": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "description": "Heure de début au format HH:mm",
            "example": "09:00"
          },
          "endTime": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "description": "Heure de fin au format HH:mm",
            "example": "10:00"
          },
          "location": {
            "type": "string",
            "pattern": "^[^<>]*$",
            "description": "Lieu de l'événement",
            "example": "Town hall"
          },
          "description": {
            "type": "string",
            "pattern": "^[^<>]*$",
            "description": "Description détaillée"
          },
          "assigneeIds": {
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ],
              "example": 1
            },
            "description": "Liste des identifiants des personnes assignées"
          },
          "assignees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CalendarAssignee"
            },
            "description": "Objets complets des personnes assignées (optionnel, peut être reconstruit côté front)"
          },
          "recurrence": {
            "$ref": "#/components/schemas/CalendarRecurrence"
          },
          "approvalStatus": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ],
            "description": "Statut d'approbation de l'événement"
          },
          "createdById": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ],
            "description": "Identifiant de l'utilisateur ayant créé l'événement",
            "example": 1
          },
          "canValidate": {
            "type": "boolean",
            "description": "Indique si l’utilisateur courant peut valider ou refuser cet événement"
          },
          "canEdit": {
            "type": "boolean",
            "description": "Indique si l’utilisateur courant peut modifier cet événement"
          },
          "canDelete": {
            "type": "boolean",
            "description": "Indique si l’utilisateur courant peut supprimer cet événement"
          },
          "visibleToRoles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "user",
                "responsable",
                "mayor"
              ]
            },
            "description": "Liste des rôles pouvant voir l'événement"
          }
        },
        "required": [
          "title",
          "date"
        ]
      },
...
      "UpdateCalendarEventApprovalBody": {
        "type": "object",
        "properties": {
          "approvalStatus": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ],
            "description": "Nouveau statut d'approbation de l'événement"
          }
        },
        "required": [
          "approvalStatus"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "BAD_REQUEST",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "PAYLOAD_TOO_LARGE",
                  "UNPROCESSABLE_ENTITY",
                  "TOO_MANY_REQUESTS",
                  "INTERNAL_ERROR",
                  "BAD_GATEWAY",
                  "SERVICE_UNAVAILABLE",
                  "GATEWAY_TIMEOUT"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            },
            "required": [
              "code",
              "message",
              "details"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/calendar/events": {
      "get": {
        "tags": [
          "Calendar"
        ],
        "summary": "Récupère les événements sur une plage de dates",
        "description": "Charge les événements utiles à la vue mois/semaine/jour",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2030-06-01"
            },
            "description": "Date de début au format YYYY-MM-DD"
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2030-06-30"
            },
            "description": "Date de fin au format YYYY-MM-DD"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des événements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CalendarEvent"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters: dates not in the YYYY-MM-DD format, from after to, or more than 1096 days apart",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Calendar API or Core API is unavailable or failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Calendar API, Core API or JWT_SECRET is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Calendar"
        ],
        "summary": "Crée un nouvel événement",
        "description": "Crée un événement calendrier et retourne l'objet enrichi avec son ID serveur",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCalendarEventBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Événement créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEvent"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Assignee outside the authorized scope, or refused by Calendar API",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Calendar API or Core API is unavailable or failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Calendar API, Core API or JWT_SECRET is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/calendar/events/{id}": {
      "patch": {
        "tags": [
          "Calendar"
        ],
        "summary": "Modifie un événement existant",
        "description": "Met à jour un événement identifié par son ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$",
              "example": "101"
            },
            "description": "Identifiant unique de l'événement"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCalendarEventBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Événement modifié avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEvent"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Action not allowed on this event",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Event not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Calendar API or Core API is unavailable or failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Calendar API, Core API or JWT_SECRET is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Calendar"
        ],
        "summary": "Supprime un événement",
        "description": "Supprime un événement identifié par son ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$",
              "example": "102"
            },
            "description": "Identifiant unique de l'événement"
          }
        ],
        "responses": {
          "204": {
            "description": "Événement supprimé avec succès"
          },
          "400": {
            "description": "Invalid id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Only the creator can delete the event",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Event not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Calendar API or Core API is unavailable or failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Calendar API, Core API or JWT_SECRET is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/calendar/events/{id}/approval": {
      "patch": {
        "tags": [
          "Calendar"
        ],
        "summary": "Met à jour le statut d’approbation d’un événement",
        "description": "Valide, refuse ou remet en attente un événement identifié par son ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$",
              "example": "101"
            },
            "description": "Identifiant unique de l'événement"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCalendarEventApprovalBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Statut d’approbation mis à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEvent"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Action not allowed on this event",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Event not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Calendar API or Core API is unavailable or failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Calendar API, Core API or JWT_SECRET is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}